move tests of mergeOrdered with delay error into nested test
